mxcp_generator/requirements.txt
pyyaml>=6.0
dataclasses>=0.6; python_version<'3.7' 